Refactored command structure
